Remove utils.cpp and put code in corresponding bytracker.cpp. Add namespaces
