+++ deploy/TensorRT/cpp/include/bytetrack/STrack.h
@@ -6,20 +6,22 @@
 using namespace cv;
 using namespace std;
 
+namespace bytetrack {
+
 enum TrackState { New = 0, Tracked, Lost, Removed };
 
 class STrack
 {
 public:
-	STrack(vector<float> tlwh_, float score);
-	~STrack();
+	STrack(std::vector<float> tlwh_, float score);
 
-	vector<float> static tlbr_to_tlwh(vector<float> &tlbr);
-	void static multi_predict(vector<STrack*> &stracks, byte_kalman::KalmanFilter &kalman_filter);
+	std::vector<float> static tlbr_to_tlwh(std::vector<float> &tlbr);
+	void static multi_predict(std::vector<STrack*> &stracks,
+		  byte_kalman::KalmanFilter &kalman_filter);
 	void static_tlwh();
 	void static_tlbr();
-	vector<float> tlwh_to_xyah(vector<float> tlwh_tmp);
-	vector<float> to_xyah();
+	std::vector<float> tlwh_to_xyah(std::vector<float> tlwh_tmp);
+	std::vector<float> to_xyah();
 	void mark_lost();
 	void mark_removed();
 	int next_id();
@@ -34,9 +36,9 @@ public:
 	int track_id;
 	int state;
 
-	vector<float> _tlwh;
-	vector<float> tlwh;
-	vector<float> tlbr;
+	std::vector<float> _tlwh;
+	std::vector<float> tlwh;
+	std::vector<float> tlbr;
 	int frame_id;
 	int tracklet_len;
 	int start_frame;
@@ -48,3 +50,5 @@ public:
 private:
 	byte_kalman::KalmanFilter kalman_filter;
 };
+
+}  // namespace bytetrack
