+++ deploy/TensorRT/cpp/include/bytetrack/dataType.h
@@ -5,6 +5,9 @@
 
 #include <Eigen/Core>
 #include <Eigen/Dense>
+
+namespace bytetrack {
+
 typedef Eigen::Matrix<float, 1, 4, Eigen::RowMajor> DETECTBOX;
 typedef Eigen::Matrix<float, -1, 4, Eigen::RowMajor> DETECTBOXSS;
 typedef Eigen::Matrix<float, 1, 128, Eigen::RowMajor> FEATURE;
@@ -34,3 +37,5 @@ typedef struct t {
 
 //linear_assignment:
 typedef Eigen::Matrix<float, -1, -1, Eigen::RowMajor> DYNAMICM;
+
+}  // namespace bytetrack
