+++ deploy/TensorRT/cpp/src/STrack.cpp
@@ -1,7 +1,8 @@
 #include "STrack.h"
 
-STrack::STrack(vector<float> tlwh_, float score)
-{
+namespace bytetrack {
+
+STrack::STrack(std::vector<float> tlwh_, float score) {
 	_tlwh.resize(4);
 	_tlwh.assign(tlwh_.begin(), tlwh_.end());
 
@@ -20,21 +21,16 @@ STrack::STrack(vector<float> tlwh_, float score)
 	start_frame = 0;
 }
 
-STrack::~STrack()
-{
-}
-
-void STrack::activate(byte_kalman::KalmanFilter &kalman_filter, int frame_id)
-{
+void STrack::activate(byte_kalman::KalmanFilter &kalman_filter, int frame_id) {
 	this->kalman_filter = kalman_filter;
 	this->track_id = this->next_id();
 
-	vector<float> _tlwh_tmp(4);
+	std::vector<float> _tlwh_tmp(4);
 	_tlwh_tmp[0] = this->_tlwh[0];
 	_tlwh_tmp[1] = this->_tlwh[1];
 	_tlwh_tmp[2] = this->_tlwh[2];
 	_tlwh_tmp[3] = this->_tlwh[3];
-	vector<float> xyah = tlwh_to_xyah(_tlwh_tmp);
+	std::vector<float> xyah = tlwh_to_xyah(_tlwh_tmp);
 	DETECTBOX xyah_box;
 	xyah_box[0] = xyah[0];
 	xyah_box[1] = xyah[1];
@@ -58,9 +54,8 @@ void STrack::activate(byte_kalman::KalmanFilter &kalman_filter, int frame_id)
 	this->start_frame = frame_id;
 }
 
-void STrack::re_activate(STrack &new_track, int frame_id, bool new_id)
-{
-	vector<float> xyah = tlwh_to_xyah(new_track.tlwh);
+void STrack::re_activate(STrack &new_track, int frame_id, bool new_id) {
+	std::vector<float> xyah = tlwh_to_xyah(new_track.tlwh);
 	DETECTBOX xyah_box;
 	xyah_box[0] = xyah[0];
 	xyah_box[1] = xyah[1];
@@ -82,12 +77,11 @@ void STrack::re_activate(STrack &new_track, int frame_id, bool new_id)
 		this->track_id = next_id();
 }
 
-void STrack::update(STrack &new_track, int frame_id)
-{
+void STrack::update(STrack &new_track, int frame_id) {
 	this->frame_id = frame_id;
 	this->tracklet_len++;
 
-	vector<float> xyah = tlwh_to_xyah(new_track.tlwh);
+	std::vector<float> xyah = tlwh_to_xyah(new_track.tlwh);
 	DETECTBOX xyah_box;
 	xyah_box[0] = xyah[0];
 	xyah_box[1] = xyah[1];
@@ -107,8 +101,7 @@ void STrack::update(STrack &new_track, int frame_id)
 	this->score = new_track.score;
 }
 
-void STrack::static_tlwh()
-{
+void STrack::static_tlwh() {
 	if (this->state == TrackState::New)
 	{
 		tlwh[0] = _tlwh[0];
@@ -128,30 +121,26 @@ void STrack::static_tlwh()
 	tlwh[1] -= tlwh[3] / 2;
 }
 
-void STrack::static_tlbr()
-{
+void STrack::static_tlbr() {
 	tlbr.clear();
 	tlbr.assign(tlwh.begin(), tlwh.end());
 	tlbr[2] += tlbr[0];
 	tlbr[3] += tlbr[1];
 }
 
-vector<float> STrack::tlwh_to_xyah(vector<float> tlwh_tmp)
-{
-	vector<float> tlwh_output = tlwh_tmp;
+std::vector<float> STrack::tlwh_to_xyah(std::vector<float> tlwh_tmp) {
+	std::vector<float> tlwh_output = tlwh_tmp;
 	tlwh_output[0] += tlwh_output[2] / 2;
 	tlwh_output[1] += tlwh_output[3] / 2;
 	tlwh_output[2] /= tlwh_output[3];
 	return tlwh_output;
 }
 
-vector<float> STrack::to_xyah()
-{
+std::vector<float> STrack::to_xyah() {
 	return tlwh_to_xyah(tlwh);
 }
 
-vector<float> STrack::tlbr_to_tlwh(vector<float> &tlbr)
-{
+std::vector<float> STrack::tlbr_to_tlwh(std::vector<float> &tlbr) {
 	tlbr[2] -= tlbr[0];
 	tlbr[3] -= tlbr[1];
 	return tlbr;
@@ -179,14 +168,14 @@ int STrack::end_frame()
 	return this->frame_id;
 }
 
-void STrack::multi_predict(vector<STrack*> &stracks, byte_kalman::KalmanFilter &kalman_filter)
-{
-	for (int i = 0; i < stracks.size(); i++)
-	{
+void STrack::multi_predict(std::vector<STrack*> &stracks,
+	  byte_kalman::KalmanFilter &kalman_filter) {
+	for (int i = 0; i < stracks.size(); i++) {
 		if (stracks[i]->state != TrackState::Tracked)
-		{
 			stracks[i]->mean[7] = 0;
-		}
+
 		kalman_filter.predict(stracks[i]->mean, stracks[i]->covariance);
 	}
 }
+
+}  // namespace bytetrack
